Renamed and reorganized the karate tests
