+++ src/test/java/services/users/UsersRunner.java
@@ -1,4 +1,4 @@
-package examples.users;
+package services.users;
 
 import com.intuit.karate.junit4.Karate;
 import org.junit.runner.RunWith;
